requirements.txt
discord.py
asyncpraw
python-dotenv
prawcore